requirements.txt
fastapi[all]
sqlalchemy
mysqlclient
python-dotenv
pydantic
alembic
pymysql
aiomysql
jwt
passlib